@@ -30,10 +30,10 @@
     "@types/jsonwebtoken": "^8.5.5",
     "@types/newrelic": "^7.0.2",
     "@types/uuid": "^7.0.4",
-    "@typescript-eslint/eslint-plugin": "^5.13.0",
+    "@typescript-eslint/eslint-plugin": "^5.19.0",
     "@typescript-eslint/parser": "^5.13.0",
     "cz-conventional-changelog": "^3.0.1",
-    "eslint": "^8.10.0",
+    "eslint": "^8.13.0",
     "eslint-config-google": "^0.14.0",
     "husky": "^7.0.4",
     "nyc": "^15.1.0",
@@ -60,7 +60,7 @@
     "graphql-ws": "^5.6.3",
     "ioredis": "^4.28.0",
     "jsonwebtoken": "^8.5.1",
-    "kidsloop-token-validation": "bitbucket:calmisland/kidsloop-token-validation",
+    "@kl-engineering/kidsloop-token-validation": "^1.6.0",
     "newrelic": "^8.4.0",
     "prom-client": "^14.0.1",
     "subscriptions-transport-ws": "^0.11.0",
