@@ -1,5 +1,5 @@
-declare module '@newrelic/apollo-server-plugin' {
-    import type {ApolloServerPlugin} from 'apollo-server-plugin-base';
+declare module "@newrelic/apollo-server-plugin" {
+    import type {ApolloServerPlugin} from "apollo-server-plugin-base";
 
     let NRApolloPlugin: ApolloServerPlugin;
     export default NRApolloPlugin;
